+++ packages/nmail_core/lib/views/settings/widgets/background_grid.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class BackgroundGrid extends StatelessWidget {
+  const BackgroundGrid({super.key, required this.children});
+
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: GridView.builder(
+        shrinkWrap: true,
+        physics: const NeverScrollableScrollPhysics(),
+        padding: EdgeInsets.zero,
+        itemCount: children.length,
+        gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
+          maxCrossAxisExtent: 148,
+          mainAxisExtent: 108,
+          crossAxisSpacing: 10,
+          mainAxisSpacing: 12,
+        ),
+        itemBuilder: (context, index) => children[index],
+      ),
+    );
+  }
+}